authorize roles by type

+++ Repository/Models/AppSettings.cs
@@ -3,7 +3,7 @@
     public class AppSettings
     {
         public string Secret { get; set; }
-        public int ExpireTime { get; set; }
+        public double ExpireTime { get; set; }
         public string ValidIssuer { get; set; }
         public string ValidAudience { get; set; }
     }
